+++ airflow/providers/common/sql/CHANGELOG.rst
@@ -24,6 +24,11 @@
 Changelog
 ---------
 
+1.1.0
+.....
+
+
+
 1.0.0
 .....
