+++ airflow/providers/qubole/provider.yaml
@@ -36,7 +36,7 @@ versions:
 
 dependencies:
   - apache-airflow>=2.2.0
-  - apache-airflow-providers-common-sql
+  - apache-airflow-providers-common-sql>=1.1.0
   - qds-sdk>=1.10.4
 
 integrations:
