+++ setup.py
@@ -622,8 +622,18 @@ def is_package_excluded(package: str, exclusion_list: List[str]) -> bool:
     return any(package.startswith(excluded_package) for excluded_package in exclusion_list)
 
 
+def strip_provider_limits(package: str) -> str:
+    """
+    Strips limit for providers in devel_all to account for packages that were not released yet.
+
+    :param package: package name (beginning of it)
+    :return: true if package should be excluded
+    """
+    return package.split(">=")[0] if package.startswith("apache-airflow-providers") else package
+
+
 devel_all = [
-    package
+    strip_provider_limits(package)
     for package in devel_all
     if not is_package_excluded(package=package, exclusion_list=PACKAGES_EXCLUDED_FOR_ALL)
 ]
