+++ tests/providers/qubole/operators/test_qubole_check.py
@@ -25,13 +25,12 @@ from qds_sdk.commands import HiveCommand
 
 from airflow.exceptions import AirflowException
 from airflow.models import DAG
+from airflow.providers.common.sql.operators.sql import SQLCheckOperator, SQLValueCheckOperator
 from airflow.providers.qubole.hooks.qubole import QuboleHook
 from airflow.providers.qubole.hooks.qubole_check import QuboleCheckHook
 from airflow.providers.qubole.operators.qubole_check import (
     QuboleCheckOperator,
     QuboleValueCheckOperator,
-    SQLCheckOperator,
-    SQLValueCheckOperator,
     _QuboleCheckOperatorMixin,
 )
